requirements.txt
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
pydantic==2.5.0
pydantic-settings==2.1.0
telethon==1.33.1
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
python-slugify==8.0.1
pypinyin==0.50.0
jinja2==3.1.2
